Adding a test to recapitulate #70
This test uses SchemaBuilder to generate a fake schema with
only prefixes in it

We then use the generic roundtripping mechanism in the
test framework to see if we can go Schema->TSV->Schema,
retaining the prefixes

This uses a simple TSV specification with two columns:

 - prefix
 - prefix_reference

As expected, this test failed, so we included a fix here.

The bug was simply that Prefix was not considered among
the list of possible Elements (it's not really an element
in the same way Class, Slot, etc are).

To fix this we add a condition checking to see if there is
a column 'prefix', and if so, if a prefix object is passed,
then use the prefix column as the primary ID (i.e prefix_prefix)